Add non modal mode for CoreBottomSheetLayout
